#2 Lint filenames
To be more consist across the project, this change adds a plugin to eslint that enforces a javascript file naming convension. The convention is to use snake_case unless there is a default export. If there is a default export, the file should be named to match the default export.