clean: Make getRoom return the room object
Returning an object containing the room was left over from when
I initially wanted to return {room, err} but I now believe it's better
to just throw an error so that it can propagate up to network.js
where the client will be sent an error message.  In the case that
the room does exist, it should just return the room.
Principle of least suprise