Fix broken ws
The API for the native WebSocket implementation wasn't quite right. Since we switched from the ws packages we assumed the event handler were functions themselves, but they need to be set to functions. Also, the onmessage handler accepts an EVENT (not data) as an argument. So the actual data is event.data