src: Support promises for makeRoom / joinRoom
This allows the consumer of PieClient the convenience of
calling pie.makeRoom(...).then().catch(), rather than
separating the then/catch logic in the onClientPresenseChanged
and onError callbacks.

This is a little complex on the Server side, because we don't
pass any kind of request id, so all it knows is that a generic
"MakeRoom" / "JoinRoom" request succeeded or failed, it doesn't know
which one.  This is why PieClient cancels previous promises if two
JoinRoom requests are made in a row.  This shouldn't be a hinderance as
a user's tab should only be in one room at a time.