ref: Merge MakeRoom and JoinRoom
After spending some time using PieClient, I've found that it
is annoying having to constantly rewrite the logic on the
client side for checking if a room already exists, hosting it
if not, and ultimately joining.  This has to be done in more places
than I expected, for example, rejoining a server if it crashes
and reboots.

This changes makes it so that JoinRoom can create the room
if the room doesn't already exist, however this is not the default
behavior because it can be important for a client to know that
a room doesn't exist when trying to join it.  For example, if I'm
trying to join a friends room, but I mistype it, it would be a
bad user experience for it to put me in another second room that
is different from my friends room.  The desireable experience
is for it to say "this room doesn't exist" so I can figure out
that I mistyped it.

So by default it will give an error message if the room doesn't
exist, but this commit adds the makeRoomIfNonExistant boolean
as a convenience for the many places where the developer just
wants to put the client in the room, regardless of whether it
exists or not.