fix: Prevent RoomManager sharing between servers
Moving roomManager instance inside of the startServer
functions allows there to be multiple PieServer instances
in the same process that don't share the same RoomManager
instance.  This is useful for tests but probably
a good idea anyway even though I don't know why else there'd be
multiple PieServer instances in the same process at this point.