fix: Prevent sending whisper messages to self
when they are not directed to self.

In sendData, to prevent lag, we pass a clients own messages back
to themself immediately.  However, we do not want to send ALL
messages back to themselves if the message is not meant to
go back to them (such as with a Whisper message).