src: Add support for soloMode
soloMode fakes a connection with a wsPie server to allow clients to
to pretend to be in a room all by themselves.  All messages they send
will come back to them as if they were connected to a real wsPieServer.
This is useful for applications that support both single user and multi
user and are built around the wsPie API.