fix: Prevent intentional close from auto reconnect
Make connect, disconnect async so they can be awaited
so as not to cross contaminate the state.  Only one websocket
should be affecting  callbacks at a time.  I don't want an
old this.ws to call onclose after this.ws has been reassigned
which would desync PieClient's state from the state
of the current this.ws connection.