fix: Only tryReconnect for connections that
have once been successfully connected.

This resolves an issue where if a pieClient attempted
to connect to a server and couldn't, they were
forever locked out of trying to connect to other
servers.  This is because it would immediately try
to reconnect which would await forever in connect
because then it would be awaiting this.disconnect
which would never resolve because the connection
was never established in the first place.

This commit fixes that issue by only attempting
to disconnect in this.connect if there is already
a CONNECTED this.ws object.

Also it only adds tryReconnect to the 'close' event
after a successful connection.