client: fix: disconnect() also calls leaveRoom()
This is important because some code may check pie.currentRoomInfo to see
if the client is currently connected to a room and we do NOT want to
clear currentRoomInfo onClose because sometimes close happens
unintentionally and we'll need currentRoomInfo to try to reconnect.
So on intentional disconnects, currentRoomInfo should be cleared so
that it doesn't retain information from the room it was previously
connected to.
Also, currentRoominfo should be set to undefined inside of leaveRoom
regardless if there is a this.ws instance or not because in soloMode
there won't be a this.ws instance but soloMode simulates a pie
connection so currentRoomInfo should still be cleared.