fix: Remove client from their previous room
when joining a new room ONLY after the new room join is successful.
This fixes an issue where if a client tried to join the same room
in two different tabs with the same clientId, it
would reject the join but also disconnect the first connection
because it would remove them from their current room thinking they
were about to join another room.