wsPie: Do not host when auto rejoining
a room after an unexpected disconnect.
If the unexpected disconnect is caused by
a server crash (where the game state is lost)
then auto rejoining results in the users
unexpectedly entering a fresh game.  Better
to just get the error message that this room does
not exist than to automatically start a new one.
Then the client using wsPie can save game state
on disconnect to allow players to resume their game safely.