PieClient: ref:  Remove solomode_client_id
It is preferrable that a client keeps the same clientId
regardless of how many times they join multiplayer vs singleplayer
and if they save a game and load it after switching between multiplayer and
singleplayer.
Because of this I'm removing "solomode_client_id" and instead opting to
generate a uuidv4 when the PieClient is first imported.  Then, if the app using
the library has cookie consent (due to GDPR), they can save that clientID and
request that they keep it when joining a server