wsPieServer: Allow hostApp to spoof messages
as clients.  This enables it to change a message asyncronously and send it with fromClient filled
as the original client.

skipHostAppHandler payload property allows the
hostApp to emit a message that will not be processed
by itself; useful for preventing infinite loops if it
processes a client message and resends it with
extra information ( like sync state )

doNotEcho property returned from handleMessage
prevents the host from echoing the message to
any clients.

npm: @websocketpie/server-bun@1.2.0